+++ enum_name_map_arg_bloc/lib/src/enma_navigator_state_extension.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 // ignore: implementation_imports
@@ -13,15 +14,24 @@ import 'transition_builder_delegate.dart';
 extension EnmaNavigatorStateExtension on NavigatorState {
   /// * [blocValue]
   ///
-  /// To provide an existing bloc to a new page.
+  /// Used to provide an existing bloc to a new page.
   ///
   /// * [blocProviders]
   ///
-  /// To provide existing blocs to a new page.
+  /// Used to provide existing blocs to a new page.
   ///
-  /// * [preventDuplicates]
+  /// * [transition]
   ///
-  /// Prevent navigate to the same page.
+  /// Used to build the route's transitions.
+  ///
+  /// * [customTransitionBuilderDelegate]
+  ///
+  /// Used to build your own custom route's transitions.
+  ///
+  /// * [curve]
+  ///
+  /// An parametric animation easing curve, i.e. a mapping of the unit interval to
+  /// the unit interval.
   ///
   /// * [duration]
   ///
@@ -35,8 +45,21 @@ extension EnmaNavigatorStateExtension on NavigatorState {
   ///
   /// {@macro flutter.widgets.PageRoute.fullscreenDialog}
   ///
+  /// * [debugPreventDuplicates]
+  ///
+  /// Prevent (accidentally) from navigating to the same page on `debug mode`.
+  ///
   /// Return a [Future]. The Future resolves when back to previous page
   /// and the [Future]'s value is the [back] method's `result` parameter.
+  ///
+  /// The `T` type argument is the type of the return value of the route.
+  ///
+  /// See also:
+  ///
+  ///  * [Curve], the interface implemented by the constants available from the
+  ///    [Curves] class.
+  ///  * [Curves], a collection of common animation easing curves.
+  @optionalTypeArgs
   Future<T?> toPage<T extends Object?, B extends BlocBase<Object?>>(
     Enum page, {
     Map<String, dynamic>? arguments,
@@ -45,20 +68,28 @@ extension EnmaNavigatorStateExtension on NavigatorState {
     RouteTransition? transition,
     TransitionBuilderDelegate? customTransitionBuilderDelegate,
     Curve? curve,
-    bool? preventDuplicates,
     Duration? duration,
     bool? opaque,
     bool? fullscreenDialog,
+    bool? debugPreventDuplicates,
   }) {
-    checkRouteType(page);
+    assert(debugAssertRouteTypeIsValid(page));
 
     final RouteConfig routeConfig = getRouteConfig(page);
 
-    if ((preventDuplicates ?? AppConfig.shouldPreventDuplicates) &&
-        widget.pages.isNotEmpty &&
-        widget.pages.last.name == page.name) {
-      duplicatedPage(page.name.runtimeType.toString());
-    }
+    assert(() {
+      final String name = effectiveRouteNameBuilder(page);
+      late final String runtimeType = objectRuntimeType(page, 'String');
+
+      if ((debugPreventDuplicates ??
+              routeConfig.debugPreventDuplicates ??
+              AppConfig.shouldPreventDuplicates) &&
+          getCurrentRouteName() == name) {
+        throw StateError("Duplicated Page: $runtimeType");
+      }
+
+      return true;
+    }());
 
     return push<T>(
       createRoute(
@@ -83,21 +114,64 @@ extension EnmaNavigatorStateExtension on NavigatorState {
     );
   }
 
-  /// [blocValue]: To provide an existing bloc to a new page.
+  /// * [blocValue]
+  ///
+  /// Used to provide an existing bloc to a new page.
   ///
   /// __Warning__: `SAUT` won't automatically handle closing the bloc.
+  ///
+  /// * [blocProviders]
+  ///
+  /// Used to provide existing blocs to a new page.
+  ///
+  /// * [result]
+  ///
+  /// If non-null, `result` will be used as the result of the route that is removed;
+  /// the future that had been returned from pushing that old route
+  /// will complete with result. The type of `result`, if provided,
+  /// must match the type argument of the class of the old route (`TO`).
+  ///
+  /// * [transition]
+  ///
+  /// Used to build the route's transitions.
+  ///
+  /// * [customTransitionBuilderDelegate]
+  ///
+  /// Used to build your own custom route's transitions.
+  ///
+  /// * [curve]
+  ///
+  /// An parametric animation easing curve, i.e. a mapping of the unit interval to
+  /// the unit interval.
   ///
   /// * [duration]
   ///
   /// The duration the transition going forwards.
   ///
-  /// Return a [Future]. The Future resolves when back to previous page
-  /// and the [Future]'s value is the [back] method's `result` parameter.
-  Future<T?>? replaceWithPage<T extends Object?, B extends BlocBase<Object?>>(
+  /// * [opaque]
+  ///
+  /// {@macro flutter.widgets.TransitionRoute.opaque}
+  ///
+  /// * [fullscreenDialog]
+  ///
+  /// {@macro flutter.widgets.PageRoute.fullscreenDialog}
+  ///
+  /// The `T` type argument is the type of the return value of the new route,
+  /// and `TO` is the type of the return value of the old route.
+  ///
+  /// See also:
+  ///
+  ///  * [Curve], the interface implemented by the constants available from the
+  ///    [Curves] class.
+  ///  * [Curves], a collection of common animation easing curves.
+  @optionalTypeArgs
+  Future<T?>? replaceWithPage<T extends Object?, B extends BlocBase<Object?>,
+      TO extends Object?>(
     Enum page, {
     Map<String, dynamic>? arguments,
     B? blocValue,
     List<BlocProviderSingleChildWidget>? blocProviders,
+    TO? result,
     RouteTransition? transition,
     TransitionBuilderDelegate? customTransitionBuilderDelegate,
     Curve? curve,
@@ -105,7 +179,7 @@ extension EnmaNavigatorStateExtension on NavigatorState {
     bool? opaque,
     bool? fullscreenDialog,
   }) {
-    checkRouteType(page);
+    assert(debugAssertRouteTypeIsValid(page));
 
     final RouteConfig routeConfig = getRouteConfig(page);
 
@@ -129,19 +203,54 @@ extension EnmaNavigatorStateExtension on NavigatorState {
         opaque: opaque ?? routeConfig.opaque,
         fullscreenDialog: fullscreenDialog ?? routeConfig.fullscreenDialog,
       ),
+      result: result,
     );
   }
 
+  /// [predicate]
+  ///
+  /// To remove routes until a route with a certain name,
+  /// use the [RoutePredicate] returned from [Saut.getModalRoutePredicate].
+  ///
+  /// To remove all the routes the replaced route, simply let [predicate] null.
+  ///
+  /// * [transition]
+  ///
+  /// Used to build the route's transitions.
+  ///
+  /// * [customTransitionBuilderDelegate]
+  ///
+  /// Used to build your own custom route's transitions.
+  ///
+  /// * [curve]
+  ///
+  /// An parametric animation easing curve, i.e. a mapping of the unit interval to
+  /// the unit interval.
+  ///
   /// * [duration]
   ///
   /// The duration the transition going forwards.
   ///
-  /// Return a [Future]. The Future resolves when back to previous page
-  /// and the [Future]'s value is the [back] method's `result` parameter.
+  /// * [opaque]
+  ///
+  /// {@macro flutter.widgets.TransitionRoute.opaque}
+  ///
+  /// * [fullscreenDialog]
+  ///
+  /// {@macro flutter.widgets.PageRoute.fullscreenDialog}
+  ///
+  /// The T type argument is the type of the return value of the new route.
+  ///
+  /// See also:
+  ///
+  ///  * [Curve], the interface implemented by the constants available from the
+  ///    [Curves] class.
+  ///  * [Curves], a collection of common animation easing curves.
+  @optionalTypeArgs
   Future<T?>?
       replaceAllWithPage<T extends Object?, B extends BlocBase<Object?>>(
     Enum page, {
-    bool Function(Route<dynamic>)? predicate,
+    RoutePredicate? predicate,
     Map<String, dynamic>? arguments,
     RouteTransition? transition,
     TransitionBuilderDelegate? customTransitionBuilderDelegate,
@@ -150,7 +259,7 @@ extension EnmaNavigatorStateExtension on NavigatorState {
     bool? opaque,
     bool? fullscreenDialog,
   }) {
-    checkRouteType(page);
+    assert(debugAssertRouteTypeIsValid(page));
 
     final RouteConfig routeConfig = getRouteConfig(page);
 
@@ -174,8 +283,16 @@ extension EnmaNavigatorStateExtension on NavigatorState {
     );
   }
 
-  void back<T>({T? result}) => pop(result);
+  /// The `T` type argument is the type of the return value of the popped route.
+  ///
+  /// The type of `result`, if provided,
+  /// must match the type argument of the class of the popped route (`T`).
+  ///
+  /// See [Navigator.pop] for more details of the semantics of popping a route.
+  @optionalTypeArgs
+  void back<T extends Object?>({T? result}) => pop(result);
 
+  /// Calls [back] repeatedly until found the page with a certain name.
   void backToPageName(String name) => popUntil((route) {
         if (route is DialogRoute) return false;
 
@@ -190,7 +307,23 @@ extension EnmaNavigatorStateExtension on NavigatorState {
                 routeName == AppConfig.initialPageName);
       });
 
-  void backToPage(dynamic page) => backToPageName(
+  /// Calls [back] repeatedly until found the page.
+  void backToPage(Enum page) => backToPageName(
         effectiveRouteNameBuilder(page),
       );
+
+  /// Trick explained here: https://github.com/flutter/flutter/issues/20451
+  /// Note `ModalRoute.of(context).settings.name` doesn't always work.
+  Route? getCurrentRoute() {
+    Route? currentRoute;
+    popUntil((route) {
+      currentRoute = route;
+      return true;
+    });
+    return currentRoute;
+  }
+
+  /// Trick explained here: https://github.com/flutter/flutter/issues/20451
+  /// Note `ModalRoute.of(context).settings.name` doesn't always work.
+  String? getCurrentRouteName() => getCurrentRoute()!.settings.name;
 }